Time: 0 ms (100.00%) | Memory: 71.7 MB (80.78%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.cpp
@@ -3,8 +3,7 @@ public:
     int maxSubArray(const vector<int>& nums) { // sliding Window
         int maxSum = nums[0];
         int currSum = nums[0];
-
-        for (size_t i = 1; i < nums.size(); ++i) {
+        for (int i = 1; i < nums.size(); ++i) {
             currSum = max(nums[i], currSum + nums[i]);
             maxSum = max(maxSum, currSum);
         };
